Refactoring of the methods being tested (part 2).
- Other methods have been removed from the description of the methods under test; they are as close as possible to the description in Swagger;
- Renaming of the method names.
